refactor: move distribution's post-close trigger to an event, isolate/retry post-processing
DistributionService.closeDistribution() now publishes DistributionEndedEvent (internal-only,
never crossing a module boundary) instead of calling the post-processor service directly.
DistributionEndedEventListener reacts to it, replacing the old generic DistributionPostProcessor
interface/list (which only had one real implementation left after ReturnBoxesMailPostProcessor
moved to reporting) with direct calls to DistributionStatisticService and the renamed
MissingCostContributionService. Statistic computation and cost contributions share one
transaction with no per-step try/catch, so a failure in either rolls both back atomically; that
transaction is retried via a shared RetryTemplate bean (config.RetryConfig) before being given up
on. The subsequent DistributionClosedEvent publish (which reporting reacts to) is deliberately
left out of the retry scope, to avoid reporting resending mails that already succeeded on a prior
attempt.

Every event-publish call site (closeDistribution, sendMails, and the listener's own publish) now
consistently logs and forwards failures instead of swallowing them. Fixed a latent MockK bug this
surfaced: `eventPublisher.publishEvent(any())` was silently resolving to the wrong overload
(publishEvent(ApplicationEvent) instead of publishEvent(Any)), making failure-injection tests
false-positive; pinned with explicit type arguments everywhere this pattern appears.

Also renamed both event listeners to the `<EventName>EventListener` convention
(DistributionEndedEventListener, DistributionClosedEventListener), and refreshed the
distribution/reporting/settings READMEs, which still described the pre-#2844 architecture in
several places.
