+++ backend/src/main/kotlin/at/wrk/tafel/admin/backend/modules/distribution/internal/statistic/MissingCostContributionService.kt
@@ -1,7 +1,6 @@
-package at.wrk.tafel.admin.backend.modules.distribution.internal.postprocessors
+package at.wrk.tafel.admin.backend.modules.distribution.internal.statistic
 
 import at.wrk.tafel.admin.backend.database.model.distribution.DistributionEntity
-import at.wrk.tafel.admin.backend.database.model.distribution.DistributionStatisticEntity
 import at.wrk.tafel.admin.backend.database.model.household.HouseholdEntity
 import at.wrk.tafel.admin.backend.database.model.household.HouseholdRepository
 import at.wrk.tafel.admin.backend.database.model.staticdata.StaticValueEntity
@@ -10,31 +9,34 @@ import at.wrk.tafel.admin.backend.database.model.staticdata.StaticValueType
 import at.wrk.tafel.admin.backend.modules.base.exception.TafelValidationException
 import org.slf4j.LoggerFactory
 import org.springframework.data.repository.findByIdOrNull
-import org.springframework.stereotype.Component
+import org.springframework.stereotype.Service
 import java.math.BigDecimal
 import java.time.LocalDate
 
 /**
- * Unlike its sibling [DistributionPostProcessor]s, which only send mail, this one mutates
- * persistent household state as a side effect of closing a distribution: for every household
- * whose [at.wrk.tafel.admin.backend.database.model.distribution.DistributionHouseholdEntity.costContributionPaid]
- * is false, it adds the current [StaticValueType.COST_CONTRIBUTION] amount to
- * [HouseholdEntity.pendingCostContribution] so it can be collected next time. Deliberately not
- * re-run by [at.wrk.tafel.admin.backend.modules.distribution.internal.DistributionService.sendMails]
- * (the manual mail re-send), since re-running it would double-count pending contributions for
- * households that already had them added when the distribution originally closed.
+ * Called by [at.wrk.tafel.admin.backend.modules.distribution.internal.DistributionEndedEventListener]
+ * right after a distribution closes. Mutates persistent household state as a side effect: for every
+ * household whose
+ * [at.wrk.tafel.admin.backend.database.model.distribution.DistributionHouseholdEntity.costContributionPaid]
+ * is false, adds the current [StaticValueType.COST_CONTRIBUTION] amount to
+ * [HouseholdEntity.pendingCostContribution] so it can be collected next time. Deliberately not called
+ * by [at.wrk.tafel.admin.backend.modules.distribution.internal.DistributionService.sendMails] (the
+ * manual mail re-send), since re-running it would double-count pending contributions for households
+ * that already had them added when the distribution originally closed - that's why `sendMails()`
+ * publishes `DistributionClosedEvent` directly instead of
+ * [at.wrk.tafel.admin.backend.modules.distribution.internal.DistributionEndedEvent], which is what
+ * triggers this service in the first place.
  */
-@Component
-class MissingCostContributionPostProcessor(
+@Service
+class MissingCostContributionService(
     private val householdRepository: HouseholdRepository,
     private val staticValueRepository: StaticValueRepository,
-) : DistributionPostProcessor {
-
+) {
     companion object {
-        private val logger = LoggerFactory.getLogger(MissingCostContributionPostProcessor::class.java)
+        private val logger = LoggerFactory.getLogger(MissingCostContributionService::class.java)
     }
 
-    override fun process(distribution: DistributionEntity, statistic: DistributionStatisticEntity) {
+    fun addMissingCostContributions(distribution: DistributionEntity) {
         val costContributionValue = staticValueRepository.findSingleValueOfType(
             StaticValueType.COST_CONTRIBUTION,
             LocalDate.now(),
