+++ backend/src/test/kotlin/at/wrk/tafel/admin/backend/modules/reporting/internal/DistributionClosedEventListenerTest.kt
@@ -10,6 +10,10 @@ import at.wrk.tafel.admin.backend.modules.base.exception.TafelValidationExceptio
 import at.wrk.tafel.admin.backend.modules.distribution.DistributionClosedEvent
 import at.wrk.tafel.admin.backend.modules.distribution.internal.testDistributionHouseholdEntity1
 import at.wrk.tafel.admin.backend.modules.distribution.internal.testDistributionHouseholdEntity2
+import at.wrk.tafel.admin.backend.modules.logistics.testFoodCollectionRoute1Entity
+import at.wrk.tafel.admin.backend.modules.logistics.testFoodCollectionRoute2Entity
+import at.wrk.tafel.admin.backend.modules.logistics.testFoodCollectionRoute3Entity
+import at.wrk.tafel.admin.backend.modules.logistics.testFoodCollectionRoute4Entity
 import at.wrk.tafel.admin.backend.modules.reporting.DailyReportService
 import at.wrk.tafel.admin.backend.modules.reporting.StatisticExportFile
 import at.wrk.tafel.admin.backend.modules.reporting.StatisticExportService
@@ -31,7 +35,7 @@ import java.time.LocalDateTime
 import java.time.format.DateTimeFormatter
 
 @ExtendWith(MockKExtension::class)
-class DistributionClosedListenerTest {
+class DistributionClosedEventListenerTest {
 
     @RelaxedMockK
     private lateinit var distributionRepository: DistributionRepository
@@ -45,7 +49,7 @@ class DistributionClosedListenerTest {
     @RelaxedMockK
     private lateinit var mailSenderService: MailSenderService
 
-    private lateinit var listener: DistributionClosedListener
+    private lateinit var listener: DistributionClosedEventListener
 
     @BeforeEach
     fun beforeEach() {
@@ -57,7 +61,7 @@ class DistributionClosedListenerTest {
             .retryOn(Exception::class.java)
             .build()
 
-        listener = DistributionClosedListener(
+        listener = DistributionClosedEventListener(
             distributionRepository,
             dailyReportService,
             statisticExportService,
@@ -67,7 +71,7 @@ class DistributionClosedListenerTest {
     }
 
     @Test
-    fun `sends daily report and statistic mails`() {
+    fun `sends daily report, statistic and return boxes mails`() {
         val distributionId = 123L
         val distributionStartDate = LocalDateTime.now().minusDays(7)
         val dateFormatted = distributionStartDate.format(DateTimeFormatter.ofPattern("dd.MM.yyyy"))
@@ -84,6 +88,12 @@ class DistributionClosedListenerTest {
         every { distribution.households } returns listOf(
             testDistributionHouseholdEntity1,
             testDistributionHouseholdEntity2,
+        )
+        every { distribution.foodCollections } returns listOf(
+            testFoodCollectionRoute1Entity,
+            testFoodCollectionRoute2Entity,
+            testFoodCollectionRoute3Entity,
+            testFoodCollectionRoute4Entity,
         )
 
         every { distributionRepository.findByIdOrNull(distributionId) } returns distribution
@@ -128,18 +138,66 @@ class DistributionClosedListenerTest {
         assertThat(statisticContextSlot.captured.getVariable("distributionDate")).isEqualTo(dateFormatted)
         assertThat(statisticAttachmentSlot.captured).hasSize(1)
         assertThat(statisticAttachmentSlot.captured[0].filename).isEqualTo("file1.csv")
+
+        val returnBoxesMailSubject = "TÖ Tafel 1030 - Retourkisten vom $dateFormatted"
+        val returnBoxesContextSlot = slot<Context>()
+        verify {
+            mailSenderService.sendHtmlMail(
+                mailType = MailType.RETURN_BOXES,
+                subject = returnBoxesMailSubject,
+                attachments = emptyList(),
+                templateName = "mails/return-boxes-mail",
+                context = capture(returnBoxesContextSlot),
+            )
+        }
+        val returnBoxesContext = returnBoxesContextSlot.captured
+        assertThat(returnBoxesContext.getVariable("distributionDate")).isEqualTo(dateFormatted)
+        assertThat(returnBoxesContext.getVariable("notes")).isEqualTo(distributionNotes)
+
+        val returnBoxes = returnBoxesContext.getVariable("returnBoxes") as ReturnBoxesDataModel
+        assertThat(returnBoxes.routes).hasSize(2)
+
+        val firstReturnBox = returnBoxes.routes.first()
+        assertThat(firstReturnBox).isEqualTo(
+            ReturnBoxesRoute(
+                name = "Route 1",
+                shops = listOf(
+                    ReturnBoxesShop(
+                        name = "2 Hofer",
+                        address = "Street 1, 1234, City",
+                        returnBoxes = "4x Category 2",
+                    ),
+                ),
+            ),
+        )
+
+        val secondReturnBox = returnBoxes.routes[1]
+        assertThat(secondReturnBox).isEqualTo(
+            ReturnBoxesRoute(
+                name = "Route 2",
+                shops = listOf(
+                    ReturnBoxesShop(
+                        name = "3 Hofer 2",
+                        address = "Street 1, 1234, City",
+                        returnBoxes = "5x Category 2",
+                    ),
+                ),
+            ),
+        )
     }
 
     @Test
-    fun `skips daily report but still sends statistic mail without customers`() {
+    fun `skips daily report but still sends statistic and return boxes mails without customers`() {
         val distributionId = 123L
         val distributionStatistic = mockk<DistributionStatisticEntity>()
 
         val distribution = mockk<DistributionEntity>()
         every { distribution.id } returns distributionId
         every { distribution.startedAt } returns LocalDateTime.now()
+        every { distribution.notes } returns null
         every { distribution.statistic } returns distributionStatistic
         every { distribution.households } returns emptyList()
+        every { distribution.foodCollections } returns emptyList()
 
         every { distributionRepository.findByIdOrNull(distributionId) } returns distribution
         every { statisticExportService.exportStatisticFiles(distributionStatistic) } returns emptyList()
@@ -148,6 +206,7 @@ class DistributionClosedListenerTest {
 
         verify(exactly = 0) { dailyReportService.generateDailyReportPdf(any()) }
         verify { statisticExportService.exportStatisticFiles(distributionStatistic) }
+        verify { mailSenderService.sendHtmlMail(mailType = MailType.RETURN_BOXES, subject = any(), attachments = any(), templateName = any(), context = any()) }
     }
 
     @Test
@@ -176,10 +235,11 @@ class DistributionClosedListenerTest {
         verify(exactly = 3) { dailyReportService.generateDailyReportPdf(distributionStatistic) }
         verify { mailSenderService.sendHtmlMail(MailType.DAILY_REPORT, any(), any(), any(), any()) }
         verify { statisticExportService.exportStatisticFiles(distributionStatistic) }
+        verify { mailSenderService.sendHtmlMail(mailType = MailType.RETURN_BOXES, subject = any(), attachments = any(), templateName = any(), context = any()) }
     }
 
     @Test
-    fun `still sends statistic mail after daily report mail exhausts all retries, then rethrows`() {
+    fun `still sends statistic and return boxes mail after daily report mail exhausts all retries, then rethrows`() {
         val (distributionId, distribution, distributionStatistic) = setupDistribution()
 
         every { dailyReportService.generateDailyReportPdf(distributionStatistic) } throws
@@ -196,10 +256,11 @@ class DistributionClosedListenerTest {
         verify(exactly = 3) { dailyReportService.generateDailyReportPdf(distributionStatistic) }
         verify { statisticExportService.exportStatisticFiles(distributionStatistic) }
         verify { mailSenderService.sendHtmlMail(mailType = MailType.STATISTICS, subject = any(), attachments = any(), templateName = any(), context = any()) }
+        verify { mailSenderService.sendHtmlMail(mailType = MailType.RETURN_BOXES, subject = any(), attachments = any(), templateName = any(), context = any()) }
     }
 
     @Test
-    fun `still attempts daily report mail after statistic mail exhausts all retries, then rethrows`() {
+    fun `still attempts daily report and return boxes mail after statistic mail exhausts all retries, then rethrows`() {
         val (distributionId, distribution, distributionStatistic) = setupDistribution()
 
         every { dailyReportService.generateDailyReportPdf(distributionStatistic) } returns ByteArray(10)
@@ -213,26 +274,50 @@ class DistributionClosedListenerTest {
         assertThat(exception.message).isEqualTo("permanent failure")
         verify(exactly = 3) { statisticExportService.exportStatisticFiles(distributionStatistic) }
         verify { mailSenderService.sendHtmlMail(MailType.DAILY_REPORT, any(), any(), any(), any()) }
+        verify { mailSenderService.sendHtmlMail(mailType = MailType.RETURN_BOXES, subject = any(), attachments = any(), templateName = any(), context = any()) }
     }
 
     @Test
-    fun `rethrows first failure with second attached as suppressed when both mails fail after retries`() {
+    fun `still attempts daily report and statistic mail after return boxes mail exhausts all retries, then rethrows`() {
+        val (distributionId, distribution, distributionStatistic) = setupDistribution()
+
+        every { dailyReportService.generateDailyReportPdf(distributionStatistic) } returns ByteArray(10)
+        every { statisticExportService.exportStatisticFiles(distributionStatistic) } returns emptyList()
+        every { mailSenderService.sendHtmlMail(MailType.RETURN_BOXES, any(), any(), any(), any()) } throws
+            TafelValidationException("permanent failure")
+
+        val exception = assertThrows<TafelValidationException> {
+            listener.onDistributionClosed(DistributionClosedEvent(distributionId))
+        }
+
+        assertThat(exception.message).isEqualTo("permanent failure")
+        verify(exactly = 3) { mailSenderService.sendHtmlMail(MailType.RETURN_BOXES, any(), any(), any(), any()) }
+        verify { mailSenderService.sendHtmlMail(MailType.DAILY_REPORT, any(), any(), any(), any()) }
+        verify { mailSenderService.sendHtmlMail(mailType = MailType.STATISTICS, subject = any(), attachments = any(), templateName = any(), context = any()) }
+    }
+
+    @Test
+    fun `rethrows first failure with others attached as suppressed when all mails fail after retries`() {
         val (distributionId, distribution, distributionStatistic) = setupDistribution()
 
         every { dailyReportService.generateDailyReportPdf(distributionStatistic) } throws
             TafelValidationException("daily report failure")
         every { statisticExportService.exportStatisticFiles(distributionStatistic) } throws
             TafelValidationException("statistic failure")
+        every { mailSenderService.sendHtmlMail(MailType.RETURN_BOXES, any(), any(), any(), any()) } throws
+            TafelValidationException("return boxes failure")
 
         val exception = assertThrows<TafelValidationException> {
             listener.onDistributionClosed(DistributionClosedEvent(distributionId))
         }
 
         assertThat(exception.message).isEqualTo("daily report failure")
-        assertThat(exception.suppressed).hasSize(1)
+        assertThat(exception.suppressed).hasSize(2)
         assertThat(exception.suppressed[0].message).isEqualTo("statistic failure")
+        assertThat(exception.suppressed[1].message).isEqualTo("return boxes failure")
         verify(exactly = 3) { dailyReportService.generateDailyReportPdf(distributionStatistic) }
         verify(exactly = 3) { statisticExportService.exportStatisticFiles(distributionStatistic) }
+        verify(exactly = 3) { mailSenderService.sendHtmlMail(MailType.RETURN_BOXES, any(), any(), any(), any()) }
     }
 
     private fun setupDistribution(): Triple<Long, DistributionEntity, DistributionStatisticEntity> {
@@ -248,6 +333,7 @@ class DistributionClosedListenerTest {
             testDistributionHouseholdEntity1,
             testDistributionHouseholdEntity2,
         )
+        every { distribution.foodCollections } returns emptyList()
 
         every { distributionRepository.findByIdOrNull(distributionId) } returns distribution
